server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.16.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0 